+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Self-hosted etcd: if `etcd-hosts.checkpoint` file exists under `${datadir}/`,
   etcd pod will restore the hosts mapping from it before etcd bootstraps.
+- Add static TLS support for self-hosted etcd mode.
 
 ### Changed